- Integrated Cobra to flesh out the cli options

+++ cmd/clean.go
@@ -1,4 +1,25 @@
-package main
+/*
+Copyright © 2022 Clive Walkden <clivewalkden@gmail.com>
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+*/
+package cmd
 
 import (
 	"context"
@@ -7,12 +28,26 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/spf13/viper"
 	"log"
-	"os"
 	"time"
 	"wasabiCleanup/internal/client/wasabi"
-	wasabiConfig "wasabiCleanup/internal/config"
+	"wasabiCleanup/internal/config"
 	"wasabiCleanup/internal/reporting"
 	"wasabiCleanup/internal/utils"
+
+	"github.com/spf13/cobra"
+)
+
+var (
+	dryRun bool
+
+	// cleanCmd represents the clean command
+	cleanCmd = &cobra.Command{
+		Use:   "clean",
+		Short: "Clean up the outdated files.",
+		Run: func(cmd *cobra.Command, args []string) {
+			clean(cmd)
+		},
+	}
 )
 
 type S3Object struct {
@@ -26,16 +61,17 @@ type S3Objects struct {
 	Size  int64
 }
 
-func main() {
-	appConfig := wasabiConfig.InitConfig()
+func init() {
+	cleanCmd.Flags().BoolVarP(&dryRun, "dryrun", "n", false, "Show what will be deleted but don't delete it")
+}
 
-	if (appConfig.Connection == wasabiConfig.S3Connection{}) {
-		log.Fatal("Sorry configuration is incomplete and can't connect to the S3 instance.")
-		os.Exit(1)
-	}
+func clean(cmd *cobra.Command) {
+	dryRun, _ := cmd.Flags().GetBool("dryrun")
+	verbose, _ := cmd.Flags().GetBool("verbose")
 
 	client := wasabi.Client()
-	report := reporting.Report{}
+
+	report := reporting.Report{DryRun: dryRun}
 
 	buckets, err := client.ListBuckets(context.TODO(), &s3.ListBucketsInput{})
 	if err != nil {
@@ -44,11 +80,11 @@ func main() {
 
 	log.Println("Working...")
 	for _, object := range buckets.Buckets {
-		if viper.GetBool("verbose") {
+		if verbose {
 			fmt.Printf("Checking Bucket %s\n", *object.Name)
 		}
 
-		if appConfig.Buckets[*object.Name] == 0 {
+		if config.AppConfig().Buckets[*object.Name] == 0 {
 			if viper.GetBool("verbose") {
 				fmt.Printf("\t- Bucket not in config, skipping\n")
 			}
@@ -70,8 +106,8 @@ func main() {
 		})
 
 		// The date we need to delete items prior to
-		comparisonDate := time.Now().AddDate(0, 0, -appConfig.Buckets[*object.Name]-1)
-		if viper.GetBool("verbose") {
+		comparisonDate := time.Now().AddDate(0, 0, -config.AppConfig().Buckets[*object.Name]-1)
+		if verbose {
 			fmt.Printf("\t- Checking files date is before %s\n", comparisonDate)
 		}
 
@@ -87,7 +123,7 @@ func main() {
 				log.Fatalf("\t\tfailed to get page %v, %v", i, err)
 			}
 
-			if viper.GetBool("verbose") {
+			if verbose {
 				fmt.Printf("\t\t- Next page (%d)\n", i)
 			}
 
@@ -98,16 +134,25 @@ func main() {
 						Key: obj.Key,
 					})
 					objectList.Size += obj.Size
-					if viper.GetBool("verbose") {
-						fmt.Printf("\t\t\t- Deleting object %s\n", *obj.Key)
-					}
-					_, err = client.DeleteObject(context.Background(), &s3.DeleteObjectInput{
-						Bucket: object.Name,
-						Key:    obj.Key,
-					})
 
-					if err != nil {
-						panic("Couldn't delete items")
+					if dryRun {
+						if verbose {
+							fmt.Printf("\t\t\t- Deleting object %s\n", *obj.Key)
+						} else {
+							fmt.Printf("\t- Deleting object %s\n", *obj.Key)
+						}
+					} else {
+						if verbose {
+							fmt.Printf("\t\t\t- Deleting object %s\n", *obj.Key)
+						}
+						_, err = client.DeleteObject(context.Background(), &s3.DeleteObjectInput{
+							Bucket: object.Name,
+							Key:    obj.Key,
+						})
+
+						if err != nil {
+							panic("Couldn't delete items")
+						}
 					}
 				} else {
 					safeList.Items = append(safeList.Items, types.ObjectIdentifier{
